+++ commands/funny/punch.py
@@ -15,9 +15,11 @@ class PunchClass(commands.Cog):
     @commands.cooldown(1, 5.0, commands.BucketType.user)
     @commands.check(lambda ctx: Database.is_registered(ctx, ctx))
     @commands.command(name='punch', aliases=['soco', 'murro'])
-    async def punch(self, ctx):
+    async def punch(self, ctx, member: discord.Member = None):
         """Comando de gifs de soco
         Use ash punch <@usuario a sua escolha>"""
+        if member is None:
+            return await ctx.send("<:alert:739251822920728708>│``Você precisa mencionar alguem!``")
         try:
             await ctx.message.delete()
         except discord.errors.Forbidden:
@@ -37,10 +39,10 @@ class PunchClass(commands.Cog):
 
             chance = randint(1, 100)
 
-            if ctx.message.mentions[0].id == self.bot.owner_id:
+            if member.id == self.bot.owner_id:
                 chance = 1
 
-            if ctx.message.mentions[0].id == self.bot.user.id:
+            if member.id == self.bot.user.id:
                 return await ctx.send('<:pqp:530031187331121152>│``Você quer me bater com meu proprio recurso?``')
 
             if chance <= 10:
@@ -62,7 +64,7 @@ class PunchClass(commands.Cog):
                 punch = 'https://i.makeagif.com/media/4-09-2016/E9n3n4.gif'
 
             punchemb = discord.Embed(title='Soco :boxing_glove: ',
-                                     description=f'**{ctx.message.mentions[0].name}** {text} **{ctx.author.name}**! '
+                                     description=f'**{member.name}** {text} **{ctx.author.name}**! '
                                                  f'{end}',
                                      color=self.color)
             punchemb.set_image(url=punch)
